remove pricereceiver endpoint and move to ChargeCommandReceiver
