+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Program.cs
@@ -53,7 +53,6 @@ namespace GreenEnergyHub.Charges.FunctionHost
             // Charges
             ChargeIngestionConfiguration.ConfigureServices(serviceCollection);
             ChargeCommandReceiverConfiguration.ConfigureServices(serviceCollection);
-            PriceCommandReceiverConfiguration.ConfigureServices(serviceCollection);
             ChargeConfirmationDataAvailableNotifierEndpointConfiguration.ConfigureServices(serviceCollection);
             ChargeRejectionDataAvailableNotifierEndpointConfiguration.ConfigureServices(serviceCollection);
             ChargeDataAvailableNotifierConfiguration.ConfigureServices(serviceCollection);
